Update README.md
`x30` gets overwritten every time a function call is made. If `main()`
made no function calls itself, `x30` would not have to be backed up.
However, this `main()` does make function calls (to `puts()`).

If we don't *save* `x30` on the stack when `main` initially enters, our
ability to properly return to whoever called `main` would be broken by
the function call to `puts()`. In all likelihood when this program ended
it would cause a crash.

These two paras imply that saving x30 is needed when main calls something as well as when it is called. Which one is true if not both?